Move TypeConstruction::*Context to TypeInference::Context::…
