+++ lib/steep.rb
@@ -60,6 +60,7 @@ require "steep/annotation_parser"
 require "steep/typing"
 require "steep/errors"
 require "steep/type_construction"
+require "steep/type_inference/context"
 require "steep/type_inference/send_args"
 require "steep/type_inference/block_params"
 require "steep/type_inference/constant_env"
